+++ src/components/input/lifecycle/lifecycle.js
@@ -1,24 +1,11 @@
-const STATE_KEYS_TO_DERIVE = [
-  'display',
-  'disabled',
-  'label',
-  'flag'
-]
-
-const STATE_KEYS_TO_UPDATE = STATE_KEYS_TO_DERIVE.concat([
-  'mouseOver',
-  'focus',
-  'decorator'
-])
-
 export function componentDidMount() {
-  this.input = document.getElementById(this.id)
   this.decorator()
 }
 
 export function getDerivedStateFromProps(nextProps, prevState) {
   let returnObject = {}
-  STATE_KEYS_TO_DERIVE.map((element) => {
+  
+  prevState.stateKeysToDerive.map((element) => {
     if (JSON.stringify(nextProps.manager(element)) !== JSON.stringify(prevState[element])) {
       returnObject[element] = nextProps.manager(element)
     }
@@ -28,7 +15,7 @@ export function getDerivedStateFromProps(nextProps, prevState) {
 
 export function shouldComponentUpdate(_, nextState) {
   let outcome = false
-  STATE_KEYS_TO_UPDATE.map((element) => {
+  this.stateKeysToUpdate.map((element) => {
     if (JSON.stringify(this.state[element]) !== JSON.stringify(nextState[element])) {
       outcome = true
     }