+++ src/components/select/select.js
@@ -1,51 +1,26 @@
 import React from 'react'
-import * as lifecycle from './lifecycle/lifecycle'
+import ManagedInput from '../input/input'
 import * as handlers from './handlers/handlers'
-import { decorator } from './decorator/decorator'
 
-export default class ManagedSelect extends React.Component {
+export default class ManagedSelect extends ManagedInput {
   constructor(props) {
     super(props)
-    this.id = this.props.manager('id')
-    this.classNames = this.props.manager('classNames')
     this.containerClass = this.classNames.container || 'managed-input select'
-    this.inputClass = this.classNames.input || 'input'
-    this.labelClass = this.classNames.label || 'label'
     this.optionsClass = this.classNames.options || 'options'
     this.optionClass = this.classNames.option || 'option'
     this.markClass = this.classNames.mark || 'mark'
-    this.errorContainerClass = this.classNames.errorContainer || 'error-container'
-    this.valueClass = this.classNames.value || 'value'
-    this.focusClass = this.classNames.focus || 'focus'
-    this.hoverClass = this.classNames.hover || 'hover'
-    this.errorClass = this.classNames.error || 'error'
-    this.componentDidMount = lifecycle.componentDidMount
-    this.shouldComponentUpdate = lifecycle.shouldComponentUpdate
-    this.componentDidUpdate = lifecycle.componentDidUpdate
-    this.onMouseOverHandler = handlers.onMouseOverHandler.bind(this)
-    this.onClickHandler = handlers.onClickHandler.bind(this)
+    this.stateKeysToDerive = [...this.stateKeysToDerive, 'options']
+    this.stateKeysToUpdate = [...this.stateKeysToUpdate, 'options', 'preSelected']
     this.onKeyDownHandler = handlers.onKeyDownHandler.bind(this)
-    this.onFocusHandler = handlers.onFocusHandler.bind(this)
-    this.onSelectHandler = handlers.onSelectHandler.bind(this)
-    this.onMouseLeaveHandler = handlers.onMouseLeaveHandler.bind(this)
     this.onBlurHandler = handlers.onBlurHandler.bind(this)
-    this.decorator = decorator.bind(this)
+    this.onSelectHandler = handlers.onSelectHandler.bind(this)
+    this.onOptionMouseOver = handlers.onOptionMouseOver.bind(this)
     this.state = {
-      display: this.props.manager('display'),
-      disabled: this.props.manager('disabled'),
-      value: this.props.manager('value'),
-      label: this.props.manager('label'),
-      error: this.props.manager('error'),
-      options:  this.props.manager('options'),
-      flag: this.props.manager('flag'),
-      mouseOver: false,
-      focus: false,
-      decorator: ''
+      ...this.state,
+      stateKeysToDerive: this.stateKeysToDerive,
+      options: this.props.manager('options'),
+      preSelected: 0
     }
-  }
-
-  static getDerivedStateFromProps(nextProps, prevstate) {
-    return lifecycle.getDerivedStateFromProps(nextProps, prevstate)
   }
 
   render() {
@@ -57,11 +32,13 @@ export default class ManagedSelect extends React.Component {
         onMouseOver={this.onMouseOverHandler}
         onMouseLeave={this.onMouseLeaveHandler}>
           <div onClick={this.onClickHandler}>
-            <div className={this.labelClass + this.state.decorator}>
+            <div
+            style={{ animation: this.state.animation }}
+            className={this.labelClass + this.state.decorator}>
               {this.state.label}
             </div>
             <div className={this.inputClass + this.state.decorator}>
-              {this.props.manager('optionDecorate', this.state.value)}
+              {this.state.options.find(option => (option.value === this.state.value)).text}
               <div className={this.markClass + this.state.decorator} />
             </div>
           </div>
@@ -70,17 +47,19 @@ export default class ManagedSelect extends React.Component {
           <div
           className={this.optionsClass + this.state.decorator}>
             {
-            this.state.options.map((option) => (
+            this.state.options.map((option, index) => (
             <div
-            key={this.props.manager('optionKey', option)}
+            key={`key-proxy-${option.value}-${index}`}
+            onMouseOver={() => this.onOptionMouseOver(index)}
             onClick={() => this.onSelectHandler(option)}
-            className={this.optionClass + this.state.decorator}>
-              {this.props.manager('optionDecorate', option)}
+            className={this.optionClass + this.state.decorator + `${this.state.preSelected === index ? ' preselected' : ''}`}>
+              {option.text}
             </div>))
             }
           </div>
           }
-          <select 
+          <select
+          ref={this.input}
           id={this.id}
           value={this.state.value}
           readOnly={true}
@@ -90,11 +69,11 @@ export default class ManagedSelect extends React.Component {
           onKeyDown={this.onKeyDownHandler}
           onBlur={this.onBlurHandler}>
             {
-            this.state.options.map((option) => (
+            this.state.options.map((option, index) => (
             <option
-            key={this.props.manager('optionKey', option + 'originalSelect')}
-            value={this.props.manager('optionValue', option)}>
-              {this.props.manager('optionDecorate', option)}
+            key={`key-original-${option.value}-${index}`}
+            value={option.value}>
+              {option.text}
             </option>))
             }
           </select>