+++ src/components/textarea/textarea.js
@@ -1,49 +1,21 @@
 import React from 'react'
-import * as lifecycle from './lifecycle/lifecycle'
+import ManagedInput from '../input/input'
 import * as handlers from './handlers/handlers'
-import { decorator } from './decorator/decorator'
 
-export default class ManagedTextarea extends React.Component {
+export default class ManagedTextarea extends ManagedInput {
   constructor(props) {
     super(props)
-    this.id = this.props.manager('id')
-    this.controlled = this.props.manager('controlled')
-    this.classNames = this.props.manager('classNames')
     this.containerClass = this.classNames.container || 'managed-input textarea'
-    this.inputClass = this.classNames.input || 'input'
-    this.labelClass = this.classNames.label || 'label'
     this.counterClass = this.classNames.counter || 'counter'
-    this.valueClass = this.classNames.value || 'value'
-    this.focusClass = this.classNames.focus || 'focus'
-    this.hoverClass = this.classNames.hover || 'hover'
-    this.errorContainerClass = this.classNames.errorContainer || 'error-container'
-    this.errorClass = this.classNames.error || 'error'
-    this.componentDidMount = lifecycle.componentDidMount
-    this.shouldComponentUpdate = lifecycle.shouldComponentUpdate
-    this.componentDidUpdate = lifecycle.componentDidUpdate
-    this.onMouseOverHandler = handlers.onMouseOverHandler.bind(this)
-    this.onMouseLeaveHandler = handlers.onMouseLeaveHandler.bind(this)
-    this.onFocusHandler = handlers.onFocusHandler.bind(this)
-    this.onBlurHandler = handlers.onBlurHandler.bind(this)
-    this.onClickHandler = handlers.onClickHandler.bind(this)
+    this.stateKeysToDerive = [...this.stateKeysToDerive, 'counterLimit']
+    this.stateKeysToUpdate = [...this.stateKeysToUpdate, 'counterLimit', 'counter']
     this.onChangeHandler = handlers.onChangeHandler.bind(this)
-    this.decorator = decorator.bind(this)
     this.state = {
-      display: this.props.manager('display'),
-      disabled: this.props.manager('disabled'),
-      value: this.props.manager('value'),
-      label: this.props.manager('label'),
-      error: this.props.manager('error'),
+      ...this.state,
+      stateKeysToDerive: this.stateKeysToDerive,
       counterLimit: this.props.manager('counterLimit'),
-      counter: 0,
-      mouseOver: false,
-      focus: false,
-      decorator: ''
+      counter: 0
     }
-  }
-
-  static getDerivedStateFromProps(nextProps, prevstate) {
-    return lifecycle.getDerivedStateFromProps(nextProps, prevstate)
   }
 
   render() { 
@@ -54,12 +26,15 @@ export default class ManagedTextarea extends React.Component {
       onMouseOver={this.onMouseOverHandler}
       onMouseLeave={this.onMouseLeaveHandler}
       onClick={this.onClickHandler}>
-        <div className={this.labelClass + this.state.decorator}>
+        <div
+        style={{ animation: this.state.animation }}
+        className={this.labelClass + this.state.decorator}>
           {this.state.label}
         </div>
         <textarea
-        id={this.id}
+        ref={this.input}
         disabled = {this.state.disabled}
+        id={this.id}
         className={this.inputClass + this.state.decorator}
         value={this.controlled ? this.state.value : undefined}
         onFocus={this.onFocusHandler}