+++ src/index.js
@@ -1,16 +1,17 @@
 import ManagedText from './components/text/text'
-import textManagerFactory from './components/text/manager-factory/manager-factory.js'
-import ManagedSelect from './components/select/select'
-import selectManagerFactory from './components/select/manager-factory/manager-factory.js'
-import ManagedButton from './components/button/button'
-import buttonManagerFactory from './components/button/manager-factory/manager-factory.js'
-import ManagedCheckbox from './components/checkbox/checkbox'
-import checkboxManagerFactory from './components/checkbox/manager-factory/manager-factory.js'
 import ManagedTextarea from './components/textarea/textarea'
-import textareaManagerFactory from './components/textarea/manager-factory/manager-factory.js'
+import ManagedSelect from './components/select/select'
+import ManagedRadio from './components/radio/radio'
+import ManagedCheckbox from './components/checkbox/checkbox'
+import ManagedMultipleCheckbox from './components/multiple-checkbox/multiple-checkbox'
+import ManagedButton from './components/button/button'
+import managerAgent from './manager-agent/manager-agent'
 
-export { ManagedText, textManagerFactory,
-         ManagedSelect, selectManagerFactory,
-         ManagedButton, buttonManagerFactory,
-         ManagedCheckbox, checkboxManagerFactory,
-         ManagedTextarea, textareaManagerFactory }
+export { ManagedText,
+         ManagedSelect,
+         ManagedTextarea,
+         ManagedRadio,
+         ManagedButton,
+         ManagedCheckbox,
+         ManagedMultipleCheckbox,
+         managerAgent }